Rise and fall Strategy.
Binary Up/Down
The bot should be able to compound after winning.
E.g. if you start with $1 and it win, that will be 
$1.94 after compound $1.94 and it win, it will give 
$3.77 and after compound $3.77 and win it will give 
$7.32 that is 3win all together. After that it will 
start from $1 again. but if it lose, it should not
martingale. but start all over again.

The adavantage of this bot is that in some cases you
have 3 to 5 winning on a roll, it will give you
quick profit.
See the screenshot of a trade that win four times.
In a situation like that we can achieve better profit.
Thanks.